store/tikv: move test function
